+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/Seasons.java
@@ -14,14 +14,14 @@ public enum Seasons {
   SPRING(0, "desc.seasonhud.season.spring", "\uEA00", Client.DEFAULT_SPRING_COLOR, SeasonHudClient.getSpringColor(),
       Rgb.seasonMap(SeasonHudClient.getSpringColor())),
 
-  SUMMER(1, "desc.seasonhud.season.summer", "\uEA01", Client.DEFAULT_SUMMER_COLOR,
-      SeasonHudClient.getSummerColor(), Rgb.seasonMap(SeasonHudClient.getSummerColor())),
+  SUMMER(1, "desc.seasonhud.season.summer", "\uEA01", Client.DEFAULT_SUMMER_COLOR, SeasonHudClient.getSummerColor(),
+      Rgb.seasonMap(SeasonHudClient.getSummerColor())),
 
-  AUTUMN(2, "desc.seasonhud.season.autumn", "\uEA02", Client.DEFAULT_AUTUMN_COLOR,
-      SeasonHudClient.getAutumnColor(), Rgb.seasonMap(SeasonHudClient.getAutumnColor())),
+  AUTUMN(2, "desc.seasonhud.season.autumn", "\uEA02", Client.DEFAULT_AUTUMN_COLOR, SeasonHudClient.getAutumnColor(),
+      Rgb.seasonMap(SeasonHudClient.getAutumnColor())),
 
-  WINTER(3, "desc.seasonhud.season.winter", "\uEA03", Client.DEFAULT_WINTER_COLOR,
-      SeasonHudClient.getWinterColor(), Rgb.seasonMap(SeasonHudClient.getWinterColor())),
+  WINTER(3, "desc.seasonhud.season.winter", "\uEA03", Client.DEFAULT_WINTER_COLOR, SeasonHudClient.getWinterColor(),
+      Rgb.seasonMap(SeasonHudClient.getWinterColor())),
 
   DRY(4, "desc.seasonhud.season.dry", "\uEA04", Client.DEFAULT_DRY_COLOR, SeasonHudClient.getDryColor(),
       Rgb.seasonMap(SeasonHudClient.getDryColor())),
@@ -37,7 +37,7 @@ public enum Seasons {
   private final String seasonIconChar;
   private final int defaultColor;
   private final Map<String, Integer> rgbMap;
-  private final String seasonNameSerilized;
+  private final String seasonNameSerialized;
   private final Component seasonNameTranslated;
   private int seasonColor;
 
@@ -48,7 +48,7 @@ public enum Seasons {
     this.defaultColor = defaultColor;
     this.seasonColor = seasonColor;
     this.rgbMap = rgbMap;
-    this.seasonNameSerilized = this.toString().toLowerCase(Locale.ROOT);
+    this.seasonNameSerialized = this.toString().toLowerCase(Locale.ROOT);
     this.seasonNameTranslated = Common.translatedText(key);
   }
 
@@ -91,8 +91,8 @@ public enum Seasons {
     return this.rgbMap;
   }
 
-  public String getSeasonNameSerilized() {
-    return this.seasonNameSerilized;
+  public String getSeasonNameSerialized() {
+    return this.seasonNameSerialized;
   }
 
   public Component getSeasonNameTranslated() {