+++ common/src/main/java/club/iananderson/seasonhud/mixin/mapatlases/MapAtlasHudMixin.java
@@ -2,7 +2,7 @@ package club.iananderson.seasonhud.mixin.mapatlases;
 
 import club.iananderson.seasonhud.client.overlays.MapAtlasesCommon;
 import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap;
-import club.iananderson.seasonhud.util.ModIds.MinimapMods;
+import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.Minimap;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
@@ -30,16 +30,16 @@ public class MapAtlasHudMixin {
   private float globalScale;
 
   @SuppressWarnings("checkstyle:ParameterName")
-  @Inject(method = "render", at = @At(value = "INVOKE", target = "Lpepjebs/mapatlases/client/ui/MapAtlasesHUD;"
-      + "drawMapComponentBiome(Lnet/minecraft/client/gui/GuiGraphics;" + "Lnet/minecraft/client/gui/Font;"
-      + "IIIFLnet/minecraft/core/BlockPos;"
+  @Inject(remap = false, method = "render", at = @At(value = "INVOKE", target = "Lpepjebs/mapatlases/client/ui"
+      + "/MapAtlasesHUD;" + "drawMapComponentBiome(Lnet/minecraft/client/gui/GuiGraphics;"
+      + "Lnet/minecraft/client/gui/Font;" + "IIIFLnet/minecraft/core/BlockPos;"
       + "Lnet/minecraft/world/level/Level;)V", shift = At.Shift.BY, by = 2), locals = LocalCapture.CAPTURE_FAILSOFT)
   private void render(GuiGraphics graphics, float partialTick, int screenWidth, int screenHeight, CallbackInfo ci,
       ItemStack atlas, MapDataHolder activeMap, ClientLevel level, LocalPlayer player, PoseStack poseStack,
       int mapWidgetSize, Anchoring anchorLocation, int off, int x, int y, float yRot, int light, int borderSize,
       float textScaling, int textHeightOffset, int actualBgSize, Font font) {
 
-    if (CurrentMinimap.mapAtlasesLoaded() && CurrentMinimap.shouldDrawMinimapHud(MinimapMods.mapAtlasesId)) {
+    if (CurrentMinimap.mapAtlasesLoaded() && CurrentMinimap.shouldDrawMinimapHud(Minimap.MAP_ATLASES)) {
       if (MapAtlasesClientConfig.drawMinimapBiome.get()) {
         textHeightOffset += (int) (10.0F * textScaling);
       }