+++ common/src/main/java/club/iananderson/seasonhud/platform/services/MinimapHelper.java
@@ -9,4 +9,6 @@ public interface MinimapHelper {
   boolean hideMapAtlases();
 
   boolean hideJourneyMap();
+
+  boolean hideXaero();
 }