+++ common/src/main/java/club/iananderson/seasonhud/util/ModIds.java
@@ -1,8 +1,5 @@
 package club.iananderson.seasonhud.util;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class ModIds {
   public static class SeasonMods {
     public static String sereneSeasonsId = "sereneseasons";
@@ -19,10 +16,6 @@ public class ModIds {
   }
 
   public static class MinimapMods {
-    public static List<String> allMods = new ArrayList<>(
-        List.of(MinimapMods.xaeroMinimapId, MinimapMods.xaeroMinimapFairplayId, MinimapMods.journeymapId,
-            MinimapMods.xaeroMinimapFairplayId, MinimapMods.ftbChunksId, MinimapMods.mapAtlasesId));
-
     public static String xaeroMinimapId = "xaerominimap";
     public static String xaeroMinimapFairplayId = "xaerominimapfair";
     public static String journeymapId = "journeymap";